Modularize app.js
All things related to configuration of the app was moved to a different
folder. Everything related to express was moved to express.js and
everything related to the view engine was moved to viewEngine.js.
Nunjucks, being a dependency, was moved to di.js where I set up my DIC.
Now, app.js responsability is basically to start the app.